refactor: make search customer bar reusable
