+++ src/features/customers/SearchCustomerBar.jsx
@@ -1,20 +1,20 @@
+/* eslint-disable react/prop-types */
 /* eslint-disable no-unused-vars */
-import { useCallback, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { debounce } from "lodash";
-import { AutoComplete, Button, Form, Modal, Space, Tooltip } from "antd";
-import { PlusOutlined } from "@ant-design/icons";
-import { useDispatch } from "react-redux";
+import { AutoComplete } from "antd";
 import { useSearchCustomers } from "./hooks/useSearchCustomers";
-import UpdateCustomerForm from "./UpdateCustomerForm";
-import { setCustomer } from "../orders/orderSlice";
 
-function SearchCustomerBar() {
-  const [isOpenModal, setIsOpenModal] = useState(false);
-  const [addCustomerForm] = Form.useForm();
+function SearchCustomerBar({ onSelectCustomer, onClear }) {
   const [inputValue, setInputValue] = useState("");
   const [searchQuery, setSearchQuery] = useState("");
   const { searchedCustomers } = useSearchCustomers(searchQuery);
-  const dispatch = useDispatch();
+
+  useEffect(() => {
+    if (inputValue === "") {
+      onClear?.();
+    }
+  }, [inputValue, onClear]);
 
   function handleSearch(value) {
     setSearchQuery(value);
@@ -31,59 +31,22 @@ function SearchCustomerBar() {
     const selectedCustomer = searchedCustomers.find(
       (customer) => customer.customerId === value,
     );
-    dispatch(setCustomer(selectedCustomer));
+    onSelectCustomer(selectedCustomer);
     setInputValue(selectedCustomer.customerName);
     setSearchQuery("");
   }
 
-  function showModal() {
-    setIsOpenModal(true);
-  }
-
-  function handleCancel() {
-    addCustomerForm.resetFields();
-    setIsOpenModal(false);
-  }
-
   return (
-    <Space.Compact className="w-[70%]">
-      <AutoComplete
-        value={inputValue}
-        className="w-[80%]"
-        placeholder="Tìm khách hàng..."
-        allowClear
-        onChange={setInputValue}
-        onSearch={debouncedHandleSearch}
-        options={customerOptions}
-        onSelect={handleSelect}
-        onClear={() => dispatch(setCustomer(null))}
-      />
-      <Tooltip title="Thêm khách hàng" placement="bottom">
-        <Button
-          style={{ borderRadius: "0 0.375rem 0.375rem 0" }}
-          icon={<PlusOutlined onClick={showModal} />}
-        />
-      </Tooltip>
-      <Modal
-        open={isOpenModal}
-        title={<span className="text-xl">Thêm khách hàng</span>}
-        width={1000}
-        okText="Thêm"
-        destroyOnClose
-        okButtonProps={{
-          form: "updateCustomerForm",
-          htmlType: "submit",
-          className: "btn-primary",
-        }}
-        cancelText="Hủy"
-        onCancel={handleCancel}
-      >
-        <UpdateCustomerForm
-          form={addCustomerForm}
-          setIsOpenModal={setIsOpenModal}
-        />
-      </Modal>
-    </Space.Compact>
+    <AutoComplete
+      value={inputValue}
+      className="w-[60%]"
+      placeholder="Tìm khách hàng..."
+      allowClear
+      onChange={setInputValue}
+      onSearch={debouncedHandleSearch}
+      options={customerOptions}
+      onSelect={handleSelect}
+    />
   );
 }
